+++ chain/actors/builtin/market/v11.go
@@ -106,6 +106,14 @@ func (s *state11) Proposals() (DealProposals, error) {
 	return &dealProposals11{proposalArray}, nil
 }
 
+func (s *state11) PendingProposals() (PendingProposals, error) {
+	proposalCidSet, err := adt11.AsSet(s.store, s.State.PendingProposals, builtin.DefaultHamtBitwidth)
+	if err != nil {
+		return nil, err
+	}
+	return &pendingProposals11{proposalCidSet}, nil
+}
+
 func (s *state11) EscrowTable() (BalanceTable, error) {
 	bt, err := adt11.AsBalanceTable(s.store, s.State.EscrowTable)
 	if err != nil {
@@ -282,6 +290,14 @@ func (s *dealProposals11) decode(val *cbg.Deferred) (*DealProposal, error) {
 
 func (s *dealProposals11) array() adt.Array {
 	return s.Array
+}
+
+type pendingProposals11 struct {
+	*adt11.Set
+}
+
+func (s *pendingProposals11) Has(proposalCid cid.Cid) (bool, error) {
+	return s.Set.Has(abi.CidKey(proposalCid))
 }
 
 func fromV11DealProposal(v11 market11.DealProposal) (DealProposal, error) {