+++ chain/actors/builtin/market/v12.go
@@ -106,6 +106,14 @@ func (s *state12) Proposals() (DealProposals, error) {
 	return &dealProposals12{proposalArray}, nil
 }
 
+func (s *state12) PendingProposals() (PendingProposals, error) {
+	proposalCidSet, err := adt12.AsSet(s.store, s.State.PendingProposals, builtin.DefaultHamtBitwidth)
+	if err != nil {
+		return nil, err
+	}
+	return &pendingProposals12{proposalCidSet}, nil
+}
+
 func (s *state12) EscrowTable() (BalanceTable, error) {
 	bt, err := adt12.AsBalanceTable(s.store, s.State.EscrowTable)
 	if err != nil {
@@ -282,6 +290,14 @@ func (s *dealProposals12) decode(val *cbg.Deferred) (*DealProposal, error) {
 
 func (s *dealProposals12) array() adt.Array {
 	return s.Array
+}
+
+type pendingProposals12 struct {
+	*adt12.Set
+}
+
+func (s *pendingProposals12) Has(proposalCid cid.Cid) (bool, error) {
+	return s.Set.Has(abi.CidKey(proposalCid))
 }
 
 func fromV12DealProposal(v12 market12.DealProposal) (DealProposal, error) {